[#57] feat : 관리자 토큰 생성·검증 및 탈퇴 API (#64)
* [#58] refactor : 기존 API에 로그인 반영
- AdminFaqController

* chore : 주석처리 정리
- UserService todo 주석 삭제
- FaqControllerTest 주석 처리
- SignUprequestDto 에 기본 생성자 추가

* [#57] feat : 회원탈퇴 API 구현

* [#57] refactor : 탈퇴시 디바이스정보도 삭제

+++ src/main/java/mashup/backend/tich/faq/controller/AdminFaqController.java
@@ -2,6 +2,7 @@ package mashup.backend.tich.faq.controller;
 
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
+import mashup.backend.tich.exception.NoAccessException;
 import mashup.backend.tich.faq.dto.FaqResponseDto;
 import mashup.backend.tich.faq.dto.FaqSaveRequestDto;
 import mashup.backend.tich.faq.dto.FaqUpdateRequestDto;
@@ -10,6 +11,7 @@ import mashup.backend.tich.faq.service.FaqService;
 import mashup.backend.tich.user.domain.Role;
 import mashup.backend.tich.user.domain.User;
 import mashup.backend.tich.user.domain.UserRepository;
+import mashup.backend.tich.user.service.AdminUserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -23,12 +25,16 @@ import java.util.List;
 public class AdminFaqController {
 
     private final AdminFaqService adminFaqService;
+    private final AdminUserService adminUserService;
     private final FaqService faqService;
     private final UserRepository userRepository; /* 삭제 예정(관리자 확인용) */
 
+
     @ApiOperation("FAQ 목록 조회")
     @GetMapping
-    public ResponseEntity<List<FaqResponseDto>> showFaqs() {
+    public ResponseEntity<List<FaqResponseDto>> showFaqs(@RequestHeader("TICH-TOKEN") String token) {
+        if(!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
+
         List<FaqResponseDto> faqResponseDto = faqService.showFaqs();
 
         return ResponseEntity.status(HttpStatus.OK).body(faqResponseDto);
@@ -36,7 +42,9 @@ public class AdminFaqController {
 
     @ApiOperation("FAQ 상세 조회")
     @GetMapping("/{faqId}")
-    public ResponseEntity<FaqResponseDto> showFaq(@PathVariable Long faqId) {
+    public ResponseEntity<FaqResponseDto> showFaq(@RequestHeader("TICH-TOKEN") String token, @PathVariable Long faqId) {
+        if(!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
+
         FaqResponseDto faqResponseDto = faqService.showFaq(faqId);
 
         return ResponseEntity.status(HttpStatus.OK).body(faqResponseDto);
@@ -44,11 +52,9 @@ public class AdminFaqController {
 
     @ApiOperation("FAQ 등록")
     @PostMapping
-    public ResponseEntity<FaqResponseDto> saveFaq(@RequestHeader String accessToken,
+    public ResponseEntity<FaqResponseDto> saveFaq(@RequestHeader("TICH-TOKEN") String token,
                                                   @RequestBody FaqSaveRequestDto requestDto) {
-        // 임시 코드 : 추후 수정
-        User user = makeTempUser();
-        // ToDo : user check (accessToken)
+        if(!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
 
         FaqResponseDto faqResponseDto = adminFaqService.saveFaq(requestDto);
 
@@ -57,11 +63,9 @@ public class AdminFaqController {
 
     @ApiOperation("FAQ 수정")
     @PutMapping
-    public ResponseEntity<FaqResponseDto> updateFaq(@RequestHeader String accessToken,
+    public ResponseEntity<FaqResponseDto> updateFaq(@RequestHeader("TICH-TOKEN") String token,
                                                     @RequestBody FaqUpdateRequestDto requestDto) {
-        // 임시 코드 : 추후 수정
-        User user = makeTempUser();
-        // ToDo : user check (accessToken)
+        if(!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
 
         FaqResponseDto faqResponseDto = adminFaqService.updateFaq(requestDto);
 
@@ -70,31 +74,13 @@ public class AdminFaqController {
 
     @ApiOperation("FAQ 삭제")
     @DeleteMapping("/{faqId}")
-    public ResponseEntity<Void> deleteFaq(@RequestHeader String accessToken,
+    public ResponseEntity<Void> deleteFaq(@RequestHeader("TICH-TOKEN") String token,
                                           @PathVariable Long faqId) {
-        // 임시 코드 : 추후 수정
-        User user = makeTempUser();
-        // ToDo : user check (accessToken)
+        if(!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
 
         adminFaqService.deleteFaq(faqId);
 
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 
-    /* 임시 코드 : 삭제 예정 */
-    private User makeTempUser() {
-        User user;
-        try {
-            user = userRepository.findByEmail("admin")
-                    .orElseThrow(() -> new NoResultException());
-        } catch (NoResultException e) {
-            user = userRepository.save(User.builder()
-                    .name("관리자")
-                    .email("admin")
-                    .picture("temp")
-                    .role(Role.ADMIN)
-                    .build());
-        }
-        return user;
-    }
 }
